+++ autorest/adal/devicetoken_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Azure/go-autorest/autorest/mocks"
+	"github.com/drake-davis/go-autorest/autorest/mocks"
 )
 
 const (