+++ autorest/authorization.go
@@ -22,7 +22,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/Azure/go-autorest/autorest/adal"
+	"github.com/drake-davis/go-autorest/autorest/adal"
 )
 
 const (