+++ autorest/autorest_test.go
@@ -18,7 +18,7 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/Azure/go-autorest/autorest/mocks"
+	"github.com/drake-davis/go-autorest/autorest/mocks"
 )
 
 func TestResponseHasStatusCode(t *testing.T) {