+++ autorest/azure/async_test.go
@@ -24,8 +24,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Azure/go-autorest/autorest"
-	"github.com/Azure/go-autorest/autorest/mocks"
+	"github.com/drake-davis/go-autorest/autorest"
+	"github.com/drake-davis/go-autorest/autorest/mocks"
 )
 
 func TestCreateFromInvalidRequestVerb(t *testing.T) {