+++ autorest/azure/azure.go
@@ -26,7 +26,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/Azure/go-autorest/autorest"
+	"github.com/drake-davis/go-autorest/autorest"
 )
 
 const (