+++ autorest/azure/cli/token.go
@@ -26,8 +26,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/Azure/go-autorest/autorest/adal"
-	"github.com/Azure/go-autorest/autorest/date"
+	"github.com/drake-davis/go-autorest/autorest/adal"
+	"github.com/drake-davis/go-autorest/autorest/date"
 	"github.com/mitchellh/go-homedir"
 )