+++ autorest/azure/example/main.go
@@ -25,9 +25,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/Azure/go-autorest/autorest"
-	"github.com/Azure/go-autorest/autorest/adal"
-	"github.com/Azure/go-autorest/autorest/azure"
+	"github.com/drake-davis/go-autorest/autorest"
+	"github.com/drake-davis/go-autorest/autorest/adal"
+	"github.com/drake-davis/go-autorest/autorest/azure"
 	"golang.org/x/crypto/pkcs12"
 )