+++ autorest/client_test.go
@@ -28,7 +28,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Azure/go-autorest/autorest/mocks"
+	"github.com/drake-davis/go-autorest/autorest/mocks"
 )
 
 func TestLoggingInspectorWithInspection(t *testing.T) {