+++ autorest/preparer_test.go
@@ -25,7 +25,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/Azure/go-autorest/autorest/mocks"
+	"github.com/drake-davis/go-autorest/autorest/mocks"
 )
 
 // PrepareDecorators wrap and invoke a Preparer. Most often, the decorator invokes the passed